requirements.txt
Django==2.1.2
django-crispy-forms==1.7.2
django-filter==2.0.0
django-tables2==2.0.1
djangorestframework==3.9.0
djangorestframework-csv==2.1.0
djangorestframework-xml==1.3.0
djangorestframework-yaml==1.0.3
drf-renderer-xlsx==0.3.0
